requirements.txt
flask==3.1.2
google-generativeai==0.8.3
anthropic==0.39.0
python-dotenv==1.0.1
crewai>=0.95.0
crewai-tools>=0.13.0
pyyaml>=6.0
